requirements.txt
requests
openai
pyyaml
colorlog